pyproject.toml
[project]
name = "rate-identification"
version = "0.1.0"
description = "PX4 rate dynamics identification from ULog files"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyulog>=0.4.0",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "numpy>=1.24",
    "scipy>=1.11",
    "click>=8.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
identify = "rate_identification.cli:main"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["src/rate_identification"]
